fetch recent message snapshot error fixed

+++ SwiftUIFirebaseChat/MainMessages/MainMessagesViewModel.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import FirebaseFirestoreSwift
 
 class MainMessagesViewModel: ObservableObject {
     
@@ -61,17 +62,35 @@ class MainMessagesViewModel: ObservableObject {
                 }
                 
                 guard let snapshot = querySnapshot else { return }
-                snapshot.documentChanges.forEach { change in
-                    let docId = change.document.documentID
+                
+                guard let documents = (snapshot as NSObject).value(forKey: "documentChanges") as? NSArray else { return }
+                
+                documents.forEach { document in
+                    guard let object = document as? NSObject else { return }
+                    guard let docs = object.value(forKey: "document") as? NSObject else { debugPrint("document was nil"); return }
+                    guard let data = docs.value(forKey: "data") as? [String: Any] else { debugPrint("data was nil"); return }
                     
+                    let docId = docs.value(forKey: "documentID") as? String
                     if let index = self.recentMessages.firstIndex(where: { rm in
                         return rm.documentId == docId
                     }) {
                         self.recentMessages.remove(at: index)
                     }
-                        
-                    self.recentMessages.insert(.init(documentId: docId, data: change.document.data()), at: 0)
+                    self.recentMessages.insert(.init(documentId: docId ?? "", data: data), at: 0)
                 }
+                
+                
+//                snapshot.documentChanges.forEach { change in
+//                    let docId = change.document.documentID
+//                    
+//                    if let index = self.recentMessages.firstIndex(where: { rm in
+//                        return rm.documentId == docId
+//                    }) {
+//                        self.recentMessages.remove(at: index)
+//                    }
+//                        
+//                    self.recentMessages.insert(.init(documentId: docId, data: change.document.data()), at: 0)
+//                }
             }
     }
